requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI/ML
openai>=1.12.0
numpy==1.24.3
httpx>=0.24.0

# Data Processing
python-multipart==0.0.6
aiofiles==23.2.1

# Database
# sqlite3 is built-in with Python - no need to install

# Development
python-dotenv==1.0.0

# TODO: Add more dependencies as needed during development
# - Vector database client (if we choose external)
# - Text processing libraries
# - Testing frameworks
